+++ ACT.SpecialSpellTimer/ConfigPanel.CombatAnalyzer.cs
@@ -23,34 +23,6 @@
 
         private SaveFileDialog combatAnalysisCSVExportSaveFileDialog = new SaveFileDialog();
 
-        /// <summary>
-        /// ExportCSVButton Click
-        /// </summary>
-        /// <param name="sender">イベント発生元</param>
-        /// <param name="e">イベント引数</param>
-        private void ExportCSVButton_Click(object sender, EventArgs e)
-        {
-            var dialog = this.combatAnalysisCSVExportSaveFileDialog;
-            dialog.RestoreDirectory = true;
-            dialog.DefaultExt = "csv";
-            dialog.Filter = "CSV File (*.csv) | *.csv";
-            dialog.OverwritePrompt = true;
-            dialog.CreatePrompt = false;
-            dialog.Title = "Export to CSV file";
-
-            if (dialog.ShowDialog() == DialogResult.OK)
-            {
-                string filename = dialog.FileName;
-                using (var sw = new StreamWriter(filename))
-                {
-                    foreach (ListViewItem item in this.CombatLogListView.Items)
-                    {
-                        var row = item.SubItems.OfType<ListViewItem.ListViewSubItem>().Skip(1).Select(s => s.Text).ToArray();
-                        sw.WriteLine(string.Join(",", row));
-                    }
-                }
-            }
-        }
 
         /// <summary>
         /// 戦闘アナライザの有効性
