+++ ACT.SpecialSpellTimer/Utility/Translate.cs
@@ -86,6 +86,13 @@
                     foreach (System.Windows.Forms.ToolStripItem c in ((System.Windows.Forms.MenuStrip)control).Items)
                         c.Text = Get(c.Text);
                 }
+
+                // Controls may have a context menu, these are not controls but they do have Text.
+                if (control is System.Windows.Forms.ContextMenuStrip)
+                {
+                    foreach (System.Windows.Forms.ToolStripItem c in ((System.Windows.Forms.ContextMenuStrip)control).Items)
+                        c.Text = Get(c.Text);
+                }
             }
             catch (Exception ex)
             {
